Adds AssayRepository tests without the PeptideModifications

+++ src/main/java/uk/ac/ebi/pride/proteomes/db/core/api/quality/ScoreRepository.java
@@ -11,14 +11,14 @@ import java.util.List;
  */
 public interface ScoreRepository extends CrudRepository<Score, Long> {
 
-    public List<Score> findByValueGreaterThan(Integer value);
-    public List<Score> findByValueLessThan(Integer value);
+    public List<Score> findByValueGreaterThan(Double value);
+    public List<Score> findByValueLessThan(Double value);
     public List<Score> findByStarCountGreaterThanAndStarType(Integer value, StarType starType);
     public List<Score> findByStarCountLessThanAndStarType(Integer value, StarType starType);
     public List<Score> findByStarType(StarType starType);
     public List<Score> findByStarCount(Integer value);
-    public List<Score> findByValueGreaterThanAndStarCountAndStarType(Integer value, Integer startCount, StarType starType);
-    public List<Score> findByValueLessThanAndStarCountAndStarType(Integer value, Integer startCount, StarType starType);
-    public List<Score> findByValueBetweenAndStarCountAndStarType(Integer min, Integer max, Integer startCount, StarType starType);
+    public List<Score> findByValueGreaterThanAndStarCountAndStarType(Double value, Integer startCount, StarType starType);
+    public List<Score> findByValueLessThanAndStarCountAndStarType(Double value, Integer startCount, StarType starType);
+    public List<Score> findByValueBetweenAndStarCountAndStarType(Double min, Double max, Integer startCount, StarType starType);
 
 }
